Moved registry enum to common module, added support for other docker registries
