+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/ClientIT.java
@@ -21,7 +21,7 @@ import io.dockstore.common.CommonTestUtilities.TestingPostgres;
 import io.dockstore.common.TestUtility;
 import io.dockstore.webservice.DockstoreWebserviceApplication;
 import io.dockstore.webservice.DockstoreWebserviceConfiguration;
-import io.dockstore.webservice.core.Registry;
+import io.dockstore.common.Registry;
 import io.dropwizard.testing.ResourceHelpers;
 import io.dropwizard.testing.junit.DropwizardAppRule;
 import io.swagger.client.ApiException;
