+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/Tool.java
@@ -37,6 +37,7 @@ import javax.persistence.UniqueConstraint;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import io.dockstore.common.Registry;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
